+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - Ability to extract the Date from the directory path, not only filename [#198].
 - New function `site.load()` to configure how to load pages, data files and components in a single place.
+- `site.copy()` now accepts an array of file extensions to copy.
+  For example: `site.copy([".jpg", ".png"])`.
 
 ### Deprecated
 - The functions `site.loadPages()`, `site.loadAssets()`, `site.loadComponents()`.